Classify silent cron repair output as failed
